refactor(scratchpad): share the plain-text editor
The scratchpad's NSTextView bridge is the only plain-text editor in the
app that turns off every substitution, and the next screen that needs
one would otherwise copy it. Lifted out of ScratchpadView as-is and
given parameters for the three things it hardcoded.

No behavior change: the scratchpad passes the same text color, the same
container inset and the same registration hook, and the placeholder
overlay now reads its font size from the editor instead of repeating
the literal.
