+++ Sources/Vorssaint/UI/PlainTextEditor.swift
@@ -1,0 +1,98 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+// Copyright (C) 2026 Vorssaint
+
+import AppKit
+import SwiftUI
+
+/// An AppKit text view configured as a pure plain-text surface: no smart
+/// quotes or dashes, no substitutions, no rich paste, with undo. SwiftUI's
+/// editor cannot switch all of that off.
+struct PlainTextEditor: NSViewRepresentable {
+    /// Both shared with callers that overlay their own text on the editor,
+    /// so a placeholder can be positioned from the same numbers as the
+    /// first line rather than from literals that drift apart.
+    /// `lineFragmentPadding` sits inside the inset and is set on the text
+    /// container below rather than assumed, so this stays the source of
+    /// the number instead of a copy of AppKit's default.
+    static let fontSize: CGFloat = 13
+    static let lineFragmentPadding: CGFloat = 5
+
+    @Binding var text: String
+    /// Appearance is applied when the view is made, not on update, so
+    /// these are configuration rather than state: a caller that derives
+    /// one from something that changes will not see it re-applied.
+    var textColor: NSColor?
+    var textContainerInset: NSSize?
+    /// Handed the text view once, for callers that need to reach it later.
+    var onCreate: ((NSTextView) -> Void)?
+
+    init(text: Binding<String>,
+         textColor: NSColor? = nil,
+         textContainerInset: NSSize? = nil,
+         onCreate: ((NSTextView) -> Void)? = nil) {
+        self._text = text
+        self.textColor = textColor
+        self.textContainerInset = textContainerInset
+        self.onCreate = onCreate
+    }
+
+    func makeNSView(context: Context) -> NSScrollView {
+        let scroll = NSTextView.scrollableTextView()
+        scroll.drawsBackground = false
+        scroll.hasVerticalScroller = true
+        scroll.autohidesScrollers = true
+        guard let textView = scroll.documentView as? NSTextView else { return scroll }
+        textView.delegate = context.coordinator
+        textView.drawsBackground = false
+        textView.font = .systemFont(ofSize: Self.fontSize)
+        textView.textContainer?.lineFragmentPadding = Self.lineFragmentPadding
+        textView.allowsUndo = true
+        textView.isRichText = false
+        textView.importsGraphics = false
+        textView.usesFontPanel = false
+        textView.isAutomaticQuoteSubstitutionEnabled = false
+        textView.isAutomaticDashSubstitutionEnabled = false
+        textView.isAutomaticTextReplacementEnabled = false
+        textView.isAutomaticSpellingCorrectionEnabled = false
+        textView.isAutomaticLinkDetectionEnabled = false
+        textView.isAutomaticDataDetectionEnabled = false
+        textView.smartInsertDeleteEnabled = false
+        if let textColor {
+            textView.textColor = textColor
+        }
+        if let textContainerInset {
+            textView.textContainerInset = textContainerInset
+        }
+        textView.string = text
+        onCreate?(textView)
+        return scroll
+    }
+
+    func updateNSView(_ nsView: NSScrollView, context: Context) {
+        guard let textView = nsView.documentView as? NSTextView,
+              textView.string != text,
+              !textView.hasMarkedText() else { return }
+        textView.string = text
+        // Programmatic replaces (load, retention, restore) invalidate undo
+        // entries recorded against the old storage; replaying one would
+        // resurrect cleared text or throw a range exception.
+        textView.undoManager?.removeAllActions()
+    }
+
+    func makeCoordinator() -> Coordinator {
+        Coordinator(text: $text)
+    }
+
+    final class Coordinator: NSObject, NSTextViewDelegate {
+        private let text: Binding<String>
+
+        init(text: Binding<String>) {
+            self.text = text
+        }
+
+        func textDidChange(_ notification: Notification) {
+            guard let textView = notification.object as? NSTextView else { return }
+            text.wrappedValue = textView.string
+        }
+    }
+}
